gpu: remove the CPU shader interpreter, SPIR-V recompile is the only path
Delete gcn_interp and the DELTA_GPU_CSRUN escape hatch: compute dispatches
always recompile to SPIR-V and run as Vulkan compute now. Clean up the GCN
paths on top of that:

- fold the duplicated straight-line VS/PS instruction switches into the
  shared emitInst; the CFG path gains proper MRT0..7 color exports
- unify the compute ALU onto the ISA-corrected GFX7 emitVop* tables (the
  interpreter-mirrored CS table had wrong opcodes in the VOP2 0x0f-0x16
  range) and make v_mul/mad i24/u24 sign/zero-extend faithfully
- replace the cs* helper wrappers with the shared Tr helpers; refresh
  stale interpreter/GLSL-era comments in gcn_spirv, gcn_translate, CMake

Also carries the in-progress PS5 runtime bringup WIP (dce scanout-buffer
accessors, sys_umtx_op signature fix, lv2/fex/GnmDriver tweaks) and
gitignores the stray build-system/ and code/ trees.

+++ delta/kern/ps4/lv2/sys_thread.h
@@ -20,5 +20,5 @@ struct thr_param;
 int PS4ABI sys_thr_new(thr_param *p, int size);
 int PS4ABI sys_thr_self(int64_t *tid);
 int PS4ABI sys_rtprio_thread(int, uint64_t, thread_prio *);
-int PS4ABI sys_umtx_op(void *, int, uint32_t, void *, void *);
+int PS4ABI sys_umtx_op(void *, int, uint64_t, void *, void *);
 }
